requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
aiomysql==0.2.0
python-dotenv==1.1.0
python-multipart==0.0.20
itsdangerous==2.2.0
